requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
openai==1.55.3
pyyaml==6.0.2
python-multipart==0.0.18
